% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_protein.R
\name{load_protein}
\alias{load_protein}
\title{Load in protein data}
\usage{
load_protein(sequence)
}
\arguments{
\item{sequence}{A string of 1-letter amino acid sequence}
}
\value{
A parsed tibble of protein data
}
\description{
Load in protein data
}
\examples{
load_protein("MSRRP")
}
